requirements_enhanced.txt
# Enhanced RAG System Requirements
# Core FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0

# PDF Processing - PyMuPDF for speed and reliability
PyMuPDF==1.23.8
pdfplumber==0.9.0  # backup option

# LangChain ecosystem for RAG
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
langchain-experimental==0.0.47
langchain-openai==0.0.2
langchain-ollama==0.0.1

# Vector stores and embeddings
docarray==0.39.1

# Text processing and retrieval
rank-bm25==0.2.2  # For BM25 retrieval

# Scientific computing and ML (for advanced retrieval)
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2

# Enhanced text processing
nltk==3.8.1
spacy==3.7.2

# Environment and configuration
python-dotenv==1.0.0

# Monitoring and tracing
langsmith==0.0.69

# Async processing
aiofiles==23.2.0
asyncio-throttle==1.0.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional: GPU acceleration (uncomment if needed)
# torch==2.1.0
# transformers==4.35.0
# sentence-transformers==2.2.2

# Database support (optional)
# sqlalchemy==2.0.23
# psycopg2-binary==2.9.7

# Production deployment
gunicorn==21.2.0
prometheus-client==0.19.0
